add progress fav button
